services/detect-tabular/pyproject.toml
[project]
name = "aegis-detect-tabular"
version = "0.1.0"
description = "Tabular drift + fairness + calibration detection worker for Aegis"
requires-python = ">=3.13"
dependencies = [
  "aegis-shared",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.28.0",
  "pydantic>=2.10.0",
  "pandas>=2.2.0",
  "numpy>=2.1.0",
  "scipy>=1.14.0",
  "fairlearn>=0.11.0",
]

# Evidently and NannyML have transitive constraints that are incompatible
# with the rest of our workspace (NannyML pins xgboost<3.0; Evidently pulls
# kaleido 0.2.x which has no macOS-arm64 wheel). The detector here wraps a
# custom drift implementation built on scipy + fairlearn (already in the
# workspace), which is sufficient for the metrics we monitor (PSI,
# KS-test, demographic parity, equal opportunity, ECE). NannyML's CBPE
# can be wired through a separate venv at the deployment boundary if
# label-free perf-est is needed in production.

[tool.uv.sources]
aegis-shared = { workspace = true }

[dependency-groups]
dev = [
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aegis_detect_tabular"]
